+++ lib/gif_gen.py
@@ -25,7 +25,7 @@ def make_gif_from_numpy(starting_noise, ending_noise, num_points, generator, gif
     noise_vectors = smooth_transition_to_noise_vector(starting_noise, ending_noise, num_points)
     noise_vectors_v = Variable(torch.from_numpy(noise_vectors))
 
-    output_images = generator(noise_vectors_v.float()).data.cpu().numpy()
+    output_images = generator(noise_vectors_v).data.cpu().numpy()
     frame_dir = os.path.join(gif_dir, 'frames')
     os.makedirs(frame_dir, exist_ok=True)
     for (i, image) in enumerate(output_images):
